+++ vmck/utils.py
@@ -1,0 +1,13 @@
+import string
+import secrets
+
+vocabulary_64 = string.ascii_letters + string.digits + '.+'
+
+
+def random_code(length, vocabulary=vocabulary_64):
+    return ''.join(secrets.choice(vocabulary) for _ in range(length))
+
+
+def is_true(value):
+    text = (value or '').lower().strip()
+    return text in ['1', 'yes', 'true', 'on', 'enabled']
